Parallel init, Blocking, 1 adapt function, intermediate save options
Parallel init - initialization is now done in parallel. The status updates how many particles are left to be initialized.

Blocking - Blocking is now allowed. The optimizer now takes in a list of objective functions and parameter indices to update at each block. The block holds all other parameters fixed.

1 adapt function - A single adapt function is used to update the particles. The "scheme" control variable switches between the different possibilities "current", "best," and "rand.'

intermediate save options - The optimizer by default saves the current output to a Global variable "SQGDE_out." This allows the optimizer to be stopped at after any iteration and recover the current state of the optimizer. The state of the change may also be save to an ".rds" file.