Create update_optim_SQGDE
Created update_optim_SQGDE, a function that updates the traces from optim_SQGDE or update_optim_SQGDE. Useful when the criteria of the optimization was not met in the iteration limit or when the trace save file was recovered before optimization finished. Must take in the same ObjFun and getAlgoParams as optim_SQGDE. TODO: write checks and tests.